doc: use ``xx`` instead of :option:`xx` where it makes sense
When we are specifying some kind of option that we are not
documenting, we shall not use :option:`xx` as it will generate a
warning. Thus, use ``xx`` in its place (or *xx* in others) as needed.

Not all converted, as others make sense to move from xx to CONFIG_xx,
which will be handled separately.

Change-Id: I98d5e70da471184f99bb491b1fa1a3b7086019d2
Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>

Former-commit-id: 3de36d0b7ceda6994106c4c693211c16b6887998