samples: power/quark_se: Add support for Sleep states
This patch extends samples/power/quark_se application so it supports
'sleep' and 'deep sleep' power states.

According to spec, in these states, the core voltage rail is turned
off. In order to be able to continue the program execution from the
point it was before entering in Sleep states, we have to save the
execution context and restore it during system startup. The current
version of QMSI doesn't provide that feature so we implement it in
the sample application. In future, QMSI will provide this functionality
and we will remove it from application code.

Even though the _sys_soc_resume hook could be easily implemented in C,
it was implemented in assembly. During Zephyr initialization, _sys_soc_
resume hook is called before the C-runtime is initialized so implementing
this function in C doesn't look appropriate. It may pass the wrong idea
that we have the C-runtime properly initialized, which is not true.

Change-Id: Id60edeb3d33c490527b94ac0f435b7f91242dd81
Signed-off-by: Andre Guedes <andre.guedes@intel.com>
Signed-off-by: Iván Briano <ivan.briano@intel.com>

Former-commit-id: ecf0844c4b4397329fd825d2876f609ff2f76040