nios2: map all sys_write* to 32-bit to work around CPU bug
Nios II appears to have an issue with the 'stbio' instruction.

When executing this code:

   0x00400848 <+136>:	stbio	r3,0(r2)

With these registers:

  r2 0x44000c	4456460
  r3 0x3	3

The memory location (which is a memory-mapped register in the
NS16550 IP block) ends up with the value 0x103 instead of 0x3 as
expected. Before the instruction ran, the register had 0 in it.

32-bit version doesn't seem to have this problem, use that
everywhere for now. This issue has been reported to Altera.

Change-Id: I4ff0ff4cc7f9b18006d3f7a777eb292924843644
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>

Former-commit-id: 75617a63803c139f085c74226a15840dcc7c58e1