stm32_mini_a15: Remove empty board file
Use a kbuild trick to force built-in.o to get built even if there isn't
any code so we can link properly.  This is cleaner than keeping around
files that don't do anything.

We keep around board.h for now since drivers and other code might expect
it to exist.

Change-Id: I891f5130f5bfd7a07b644ee0223b18fd86061cfa
Signed-off-by: Kumar Gala <kumar.gala@linaro.org>

Former-commit-id: 5cb394bfa0a2ebaaf80e65127056d6bce2ff1b7e