olimexino_stm32: Remove empty board file
Use a kbuild trick to force built-in.o to get built even if there isn't
any code so we can link properly.  This is cleaner than keeping around
files that don't do anything.

We keep around board.h for now since drivers and other code might expect
it to exist.

Change-Id: I2e824e1baf582517e713cabe1850c9accd509a5a
Signed-off-by: Kumar Gala <kumar.gala@linaro.org>

Former-commit-id: bef89748cdfe34cb948e39cc90dc0af4f9c085cb