frdm_k64f: Remove empty board file
Use a kbuild trick to force built-in.o to get built even if there isn't
any code so we can link properly.  This is cleaner than keeping around
files that don't do anything.

We keep around board.h for now since drivers and other code might expect
it to exist.

Change-Id: Ia29cbf53670e254202c3d8916a28a1ba2254107c
Signed-off-by: Kumar Gala <kumar.gala@linaro.org>

Former-commit-id: 764ebd485af4f00fa7f68f0c5944dc5ef36d2482