Revert "Bluetooth: SMP: Workaround LE SC bug in iOS"
This reverts commit 8784da8d62cf728d5dced7e9821b8451bce9fd07 [formerly 15c1fed5b65cf51a92b270b0353a039f2512e314].

This workaround is no longer needed as it was verified that pairing
succeed without it. Also I'm no longer sure if initial pairing issue
was due to missing bit in key distirbution field or if it was due to
iOS already had device with this address paired (ie removing device
bonding from settings application fix the problem).

Change-Id: I6d69b9afb04afc83139d57d2dffe1b62d40afe5b
Signed-off-by: Szymon Janc <ext.szymon.janc@tieto.com>

Former-commit-id: b944c0bd653c5f1acde98ef53455c56cabd148c3